# *: reuse chunk in the histogram