--- a/pkg/util/chunk/BUILD.bazel
+++ b/pkg/util/chunk/BUILD.bazel
@@ -32,6 +32,7 @@ go_library(
         "//pkg/util/hack",
         "//pkg/util/logutil",
         "//pkg/util/memory",
+        "//pkg/util/syncutil",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@org_golang_x_sys//cpu",