--- a/pkg/util/chunk/chunk.go
+++ b/pkg/util/chunk/chunk.go
@@ -71,6 +71,11 @@ func NewChunkWithCapacity(fields []*types.FieldType, capacity int) *Chunk {
 	return New(fields, capacity, capacity)
 }
 
+// NewChunkFromPoolWithCapacity creates a new chunk with field types and capacity from the pool.
+func NewChunkFromPoolWithCapacity(fields []*types.FieldType, initCap int) *Chunk {
+	return getChunkFromPool(initCap, fields)
+}
+
 // New creates a new chunk.
 //
 //	cap: the limit for the max number of rows.
@@ -668,4 +673,8 @@ func (c *Chunk) AppendPartialRows(colOff int, rows []Row) {
 			appendCellByCell(dstCol, srcRow.c.columns[i], srcRow.idx)
 		}
 	}
+}
+
+func (c *Chunk) Destory(initCap int, fields []*types.FieldType) {
+	putChunkFromPool(initCap, fields, c)
 }